Moved from local to LCG json library
main/include/json.hpp was a copy of an external library. This commit
drops it in favor of using the released version of it in LCG.
It has the big advantage to be maintained and thus to be recent. The
current copy was not and thus created conflicts with the usage of the
same library inside Gaudi.