SMOG2 PV reconstruction
Adaption of the Primary Vertex reconstruction algorithm to also reconstruct SMOG2 PVs.

Within the SMOG2 project, Run3 proton-gas collisions will be limited to z in [-500, -300] mm, where the gas will be confined. In order to efficiently reconstruct SMOG2 PVs coping with their lower track multiplicity and with the event geometry with mainly high-pseudorapidity tracks, the thresholds set in the histogramming and clustering phases of the PV reconstruction need to be loosened. By asking a different threshold as a function of the z (track poca, seed or vertex position, depending on the phase), this is proved to not affect the pp performance, while most inefficiency on SMOG2 events is recovered (see enclosed plots and presentations linked there).